+++ component/admin/src/Controller/SyncController.php
@@ -1,5 +1,5 @@
 <?php
-namespace Xdecaro\Component\Competitions\Administrator\Controller;
+namespace xdecaro\Component\Competitions\Administrator\Controller;
 
 defined('_JEXEC') or die;
 
@@ -9,7 +9,7 @@ use Joomla\CMS\MVC\Controller\BaseController;
 use Joomla\CMS\Response\JsonResponse;
 use Joomla\CMS\Session\Session;
 use Joomla\Database\DatabaseInterface;
-use Xdecaro\Component\Competitions\Administrator\Helper\LiveSyncHelper;
+use xdecaro\Component\Competitions\Administrator\Helper\LiveSyncHelper;
 
 final class SyncController extends BaseController
 {